Extract extract_commands_for_help into a Tools module
